pyproject.toml
[project]
name = "backend-structure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "bcrypt>=5.0.0",
    "bson>=0.5.10",
    "edge-tts>=7.2.3",
    "fastapi>=0.121.2",
    "google-genai>=1.51.0",
    "google-search-results>=2.4.2",
    "groq>=0.35.0",
    "httpx>=0.28.1",
    "python-jose[cryptography]>=3.3.0",
    "langchain>=1.0.7",
    "langchain-core>=1.0.5",
    "langchain-groq>=1.0.1",
    "loguru>=0.7.3",
    "numpy>=2.3.5",
    "orjson>=3.11.4",
    "pandas>=2.3.3",
    "passlib>=1.7.4",
    "pre-commit>=4.4.0",
    "psutil>=7.1.3",
    "pycryptodome>=3.23.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "supabase>=2.24.0",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
    "pyaudio>=0.2.14",
]
